+++ test/html/test_html.py
@@ -11,12 +11,8 @@ from test.conftest import assert_pdf_equal
 HERE = Path(__file__).resolve().parent
 
 
-class MyFPDF(FPDF, HTMLMixin):
-    pass
-
-
 def test_html_images(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
 
     initial = 10
@@ -38,7 +34,7 @@ def test_html_images(tmp_path):
 
 
 def test_html_features(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html("<p><b>hello</b> world. i am <i>tired</i>.</p>")
     pdf.write_html("<p><u><b>hello</b> world. i am <i>tired</i>.</u></p>")
@@ -64,6 +60,7 @@ def test_html_features(tmp_path):
     pdf.write_html("<blockquote>hello blockquote</blockquote>")
     pdf.write_html("<ul><li>li1</li><li>another</li><li>l item</li></ul>")
     pdf.write_html("<ol><li>li1</li><li>another</li><li>l item</li></ol>")
+    pdf.write_html("<dl><dt>description title</dt><dd>description details</dd></dl>")
     pdf.write_html('<table width="50"></table>')
     pdf.write_html("<img></img>")
     pdf.write_html(
@@ -179,7 +176,7 @@ def test_html_features(tmp_path):
 
 
 def test_html_simple_table(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.set_font_size(30)
     pdf.add_page()
     pdf.write_html(
@@ -195,7 +192,7 @@ def test_html_simple_table(tmp_path):
 
 
 def test_html_table_line_separators(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.set_font_size(30)
     pdf.add_page()
     pdf.write_html(
@@ -212,7 +209,7 @@ def test_html_table_line_separators(tmp_path):
 
 
 def test_html_table_th_inside_tr_issue_137(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         """<table border="1">
@@ -230,7 +227,7 @@ def test_html_table_th_inside_tr_issue_137(tmp_path):
 
 
 def test_html_table_with_border(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.set_font_size(30)
     pdf.add_page()
     pdf.write_html(
@@ -246,7 +243,7 @@ def test_html_table_with_border(tmp_path):
 
 
 def test_html_bold_italic_underline(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.set_font_size(30)
     pdf.add_page()
     pdf.write_html(
@@ -264,7 +261,7 @@ def test_customize_ul(tmp_path):
             <li><b>term2</b>: definition2</li>
         </ul>"""
     # 1. Customizing through class attributes:
-    class CustomPDF(FPDF, HTMLMixin):
+    class CustomPDF(FPDF):
         li_tag_indent = 5
         ul_bullet_char = "\x86"
 
@@ -286,7 +283,7 @@ def test_customize_ul(tmp_path):
 
 
 def test_img_inside_html_table(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         """<table>
@@ -304,7 +301,7 @@ def test_img_inside_html_table(tmp_path):
 
 
 def test_img_inside_html_table_without_explicit_dimensions(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         """<table>
@@ -326,7 +323,7 @@ def test_img_inside_html_table_without_explicit_dimensions(tmp_path):
 
 
 def test_img_inside_html_table_centered(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         """<table>
@@ -344,7 +341,7 @@ def test_img_inside_html_table_centered(tmp_path):
 
 
 def test_img_inside_html_table_centered_with_align(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         """<table>
@@ -364,7 +361,7 @@ def test_img_inside_html_table_centered_with_align(tmp_path):
 
 
 def test_img_inside_html_table_centered_with_caption(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         """<table border="1">
@@ -387,7 +384,7 @@ def test_img_inside_html_table_centered_with_caption(tmp_path):
 
 
 def test_html_table_with_empty_cell_contents(tmp_path):  # issue 349
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.set_font_size(30)
     pdf.add_page()
     # Reference table cells positions:
@@ -414,7 +411,7 @@ def test_html_table_with_empty_cell_contents(tmp_path):  # issue 349
 
 
 def test_html_justify_paragraph(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         '<p align="justify">'
@@ -428,7 +425,7 @@ def test_html_justify_paragraph(tmp_path):
 
 
 def test_issue_156(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_font("Roboto", style="B", fname="test/fonts/Roboto-Bold.ttf")
     pdf.set_font("Roboto", style="B")
     pdf.add_page()
@@ -444,7 +441,7 @@ def test_issue_156(tmp_path):
 
 
 def test_html_font_color_name(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         '<font color="crimson"><p>hello in crimson</p></font>'
@@ -458,7 +455,7 @@ def test_html_font_color_name(tmp_path):
 
 
 def test_html_heading_hebrew(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_font("DejaVuSans", fname=HERE / "../fonts/DejaVuSans.ttf")
     pdf.set_font("DejaVuSans")
     pdf.add_page()
@@ -467,7 +464,7 @@ def test_html_heading_hebrew(tmp_path):
 
 
 def test_html_headings_line_height(tmp_path):  # issue-223
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     long_title = "The Quick Brown Fox Jumped Over The Lazy Dog "
     pdf.write_html(
@@ -484,7 +481,7 @@ def test_html_headings_line_height(tmp_path):  # issue-223
 
 
 def test_html_custom_heading_sizes(tmp_path):  # issue-223
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         """<h1>This is a H1</h1>
@@ -499,7 +496,7 @@ def test_html_custom_heading_sizes(tmp_path):  # issue-223
 
 
 def test_bgcolor_in_table(tmp_path):  # issue-512
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         """<table>
@@ -521,10 +518,51 @@ def test_bgcolor_in_table(tmp_path):  # issue-512
 
 
 def test_html_superscript(tmp_path):
-    pdf = MyFPDF()
+    pdf = FPDF()
     pdf.add_page()
     pdf.write_html(
         "<h1>Superscript/Subscript test</h1>"
         "2<sup>56</sup> more line text<sub>(idx)</sub>"
     )
     assert_pdf_equal(pdf, HERE / "html_superscript.pdf", tmp_path)
+
+
+def test_html_description(tmp_path):
+    pdf = FPDF()
+    pdf.add_page()
+    pdf.write_html(
+        """
+           <dt>description title</dt>
+           <dd>description details</dd>
+            <dl>
+                <dt>description title</dt>
+                <dd>description details</dd>
+            </dl>
+        """
+    )
+    assert_pdf_equal(pdf, HERE / "html_description.pdf", tmp_path)
+
+
+def test_html_HTMLMixin_deprecation_warning(tmp_path):
+    class PDF(FPDF, HTMLMixin):
+        pass
+
+    msg = (
+        "The HTMLMixin class is deprecated. "
+        "Simply use the FPDF class as a replacement."
+    )
+
+    with pytest.warns(DeprecationWarning, match=msg):
+        pdf = PDF()
+        pdf.add_page()
+        pdf.write_html(
+            """
+           <dt>description title</dt>
+           <dd>description details</dd>
+            <dl>
+                <dt>description title</dt>
+                <dd>description details</dd>
+            </dl>
+        """
+        )
+        assert_pdf_equal(pdf, HERE / "html_description.pdf", tmp_path)